requirements.txt
absl-py==0.11.0
antlr4-python3-runtime==4.8
appdirs==1.4.3
apturl==0.5.2
astunparse==1.6.2
atomicwrites==1.1.5
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
biopython==1.79
bleach==3.1.1
blinker==1.4
blosc==1.7.0
Brlapi==0.7.0
cachetools==4.0.0
caffe==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dask==2.8.1+dfsg
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fairseq==0.10.2
falkon==0.6.3
fasteners==0.14.1
filelock==3.0.12
flake8==3.7.9
flatbuffers==1.12
fsspec==0.6.1
future==0.18.2
gast==0.4.0
goatools==1.1.6
google-auth==1.5.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.29.1
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.0.12
hydra-core==1.1.0
idna==2.8
imageio==2.4.1
importlib-metadata==1.5.0
importlib-resources==5.2.2
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
jax==0.2.19
jaxlib==0.1.70+cuda111
jdcal==1.0
jedi==0.15.2
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
locket==0.2.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
msgpack==0.6.2
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numexpr==2.7.1
numpy==1.21.2
nvidia-ml-py==7.352.0
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.0
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parameterized==0.7.0
paramiko==2.6.0
parso==0.5.2
partd==1.0.0
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkg_resources==0.0.0
pluggy==0.13.0
ply==3.11
portalocker==2.3.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.17.3
psutil==5.5.1
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.7.1
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycuda==2019.1.2
pycups==1.9.73
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
pygpu==0.7.6
pyinotify==0.9.6
PyJWT==1.7.1
pykeops==1.4.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytools==2019.1.1
pytz==2019.3
PyWavelets==0.5.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
regex==2021.8.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.0.0
requests-unixsocket==0.2.0
rsa==4.0
sacrebleu==2.0.0
sacremoses==0.0.45
scikit-cuda==0.5.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.3.3
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
ssh-import-id==5.10
statsmodels==0.12.2
systemd-python==234
tables==3.6.1
tabulate==0.8.9
tensorboard==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
Theano==1.0.4
tokenizers==0.10.3
toolz==0.9.0
torch==1.9.0
torchvision==0.10.0
tornado==5.1.1
tqdm==4.62.1
traitlets==4.3.3
transformers==4.9.2
typing-extensions==3.7.4.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
wget==3.2
widgetsnbextension==2.0.0
wrapt==1.11.2
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.1
xlwt==1.3.0
zipp==3.5.0
